refactor(python): split binding tests by category
