+++ crates/bashkit-python/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Python binding test modules."""
