+++ crates/bashkit-python/tests/_bashkit_categories.py
